Sum of Products(SOP) into Products of Sum (POS)
Conversion of  Sum of Products (SOP) to Product of Sum (POS) which is commonly seen in digital circuit studies including Boolean algebra. SOP is a method of describing a Boolean expression using a set of min terms or product terms. POS is a method of describing a Boolean expression using a set of max terms or sum terms.